requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tavily-python>=0.3.0
pandas>=2.0.0
numpy>=1.24.0